Rename `bloodsugar.view` package to `bloodsugar.ui`
