+++ app/src/main/java/org/simple/clinic/summary/bloodsugar/ui/BloodSugarSummaryViewInjector.kt
@@ -1,4 +1,4 @@
-package org.simple.clinic.summary.bloodsugar.view
+package org.simple.clinic.summary.bloodsugar.ui
 
 interface BloodSugarSummaryViewInjector {
   fun inject(target: BloodSugarSummaryView)
