+++ app/src/main/res/layout/list_patientsummary_bloodsugar_measurement.xml
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
-<org.simple.clinic.summary.bloodsugar.view.BloodSugarItemView xmlns:android="http://schemas.android.com/apk/res/android"
+<org.simple.clinic.summary.bloodsugar.ui.BloodSugarItemView xmlns:android="http://schemas.android.com/apk/res/android"
   android:layout_width="match_parent"
   android:layout_height="wrap_content" />
